Enable changing activating key.
The keys [ and ] can now be changed to anything by setting
g:unimpaired_left and g:unimpaired_right. If unset, default value is
still [ and ]. Useful for e.g. scandinavian keyboards where [ and ]
are a two- or three-key chord.
If <, > or = are used for g:unimpaired_left or g:unimpaired_right. The
<[pP], >[pP] and =[pP] mappings are changed to <i[pP], >i[pP] and =i[pP]
respectively.
Those changes are undocumented.