Add support for repeating previous/next commands
This is useful for mapping it to something convenient like `üü` and `++`
on a German keyboard layout, allowing for context-sensitive movement by
default (e.g. forwarding to `[c`, `]c` in diff mode), and remembering
the last used command group.

Other plugins are able to hook into this by setting `g:unimpaired_prevnext`.

It uses `:normal` in g:unimpaired_prevnext explicitly.  This allows to
specify the whole command for `:exe`, and does not require a mapping.

The mapping is used internally to handle the `zv`, but that could be
included in the command already now.

It uses the `g:` namespace, which is important for repeating `[q/]q`
across buffers.